requirements.txt
boto3==1.16.10
epl.protobuf.v1==1.0.3
google-cloud-storage>=1.14.0
grpcio-tools==1.33.2
protobuf>=3.8.0
requests
retry
